+++ TFT/src/User/API/UI/CharIcon.h
@@ -6,6 +6,7 @@ extern "C" {
 #endif
 
 #include <stdint.h>
+#include "GUI.h"
 
 typedef enum
 {
@@ -21,6 +22,7 @@ extern const uint16_t charIconColor[CHARICON_NUM];
 
 uint8_t * IconCharSelect(uint8_t sel);
 char * IconChar(uint8_t sel);
+void drawCharIcon(const GUI_RECT * rect, ALIGN_POSITION iconalign, uint16_t iconindex, bool drawBgColor, uint16_t btn_color);
 
 #ifdef __cplusplus
 }
