+++ TFT/src/User/API/UI/GUI.h
@@ -74,6 +74,7 @@ void GUI_FillCircle(uint16_t x0, uint16_t y0, uint16_t r);
 void GUI_SetRange(int16_t x0, int16_t y0, int16_t x1, int16_t y1);
 void GUI_CancelRange(void);
 void GUI_DrawPixel(int16_t x, int16_t y, uint16_t color);
+GUI_POINT getTextStartPoint(uint16_t sx, uint16_t sy, uint16_t ex, uint16_t ey,ALIGN_POSITION pos, const char * textchar);
 void GUI_DispOne(int16_t sx, int16_t sy, const CHAR_INFO *pInfo);
 
 // display string from char array pointers
